pyproject.toml
[tool.poetry]
name = "prisma-access-decrypt-pcap"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "^3.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
